+++ .idea/misc.xml
@@ -3,5 +3,5 @@
   <component name="Black">
     <option name="sdkName" value="Python 3.13 (3.13.0)" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.10 (WishJob)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.10 (wishjob)" project-jdk-type="Python SDK" />
 </project>
