--- job/models.py
@@ -1,8 +1,0 @@
-from django.db import models
-
-# Create your models here.
-
-class Job(models.Model):
-    title = models.CharField(max_length=100)
-    description = models.TextField()
-    created_at = models.DateTimeField(auto_now_add=True)
