--- job/urls.py
@@ -1,4 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns=[]
