+++ users/models.py
@@ -1,0 +1,14 @@
+from django.contrib.auth.models import AbstractUser
+from django.db import models
+from django.core.validators import RegexValidator
+
+class CustomUser(AbstractUser):
+    phoneNum = models.CharField(
+        max_length=11,
+        unique=True,
+        validators=[RegexValidator(regex=r'^01[0-9]{8,9}$', message='Enter a valid phone number')]
+    )
+    birth = models.DateField()
+
+    def __str__(self):
+        return self.username
